feat: Move VC API endpoints to functions, to more easily create your own API server, only supporting certain endpoints
